requirements.txt
streamlit
langchain
langgraph
chromadb
langchain-community
langchain-ollama
python-dotenv 
tiktoken
langchain-google-genai
fastapi[standard]
langchain-chroma
langchain-pinecone
pinecone-client